pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "versioned-config"
version = "0.2.0"
description = "Settings and file I/O management using a configuration YAML file"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
  { name = "Nathaniel Henry", email = "nat@henryspatialanalysis.com" }
]
keywords = ["versioning", "config", "yaml", "file-io", "data-pipeline"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
requires-python = ">=3.9"
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
pandas = ["pandas>=1.3"]
geo = ["geopandas>=0.12"]
raster = ["rasterio>=1.3"]
xarray = ["xarray>=2022.3"]
dbfread = ["dbfread>=2.0"]
excel = ["openpyxl>=3.0"]
all = [
  "pandas>=1.3",
  "openpyxl>=3.0",
  "geopandas>=0.12",
  "rasterio>=1.3",
  "xarray>=2022.3",
  "dbfread>=2.0",
]
docs = [
  "sphinx>=7.0",
  "furo>=2023.9.10",
]
dev = [
  "pandas>=1.3",
  "openpyxl>=3.0",
  "pytest>=7.0",
  "pytest-cov>=4.0",
]

[project.urls]
"Homepage" = "https://github.com/henryspatialanalysis/versioning"
"Bug Tracker" = "https://github.com/henryspatialanalysis/versioning/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"versioned_config" = ["data/*.yaml", "data/*.csv"]

[tool.pytest.ini_options]
testpaths = ["tests"]
